made new spectrometer base class, moved BLE operations into new derived class (all functionality tested)
